+++ src/OpenTelemetry.Instrumentation.SqlClient/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* Added overloads which accept a name to the `TracerProviderBuilder`
+  `AddSqlClientInstrumentation` extension to allow for more fine-grained options
+  management
+  ([#XXXX](https://github.com/open-telemetry/opentelemetry-dotnet/pull/XXXX))
+
 ## 1.0.0-rc9.6
 
 Released 2022-Aug-18